+++ common/src/main/java/com/leobeliik/extremesoundmuffler/Constants.java
@@ -14,4 +14,5 @@ public class Constants {
     public static final Logger LOG = LogManager.getLogger("Extreme Sound Muffler");
     public static final KeyMapping soundMufflerKey = SoundMufflerCommon.mufflerKey();
     public static final List<Anchor> emptyAnchorList = IntStream.range(0, 10).mapToObj(i -> new Anchor(i, "Anchor " + i)).collect(Collectors.toList());
+    public static boolean isCustomSkinLoader;
 }