+++ fabric/src/main/java/com/leobeliik/extremesoundmuffler/SoundMufflerFabric.java
@@ -10,6 +10,7 @@ import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import net.fabricmc.fabric.api.client.screen.v1.ScreenEvents;
 import net.fabricmc.fabric.api.client.screen.v1.ScreenMouseEvents;
+import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.client.gui.components.events.GuiEventListener;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.inventory.CreativeModeInventoryScreen;
@@ -28,6 +29,7 @@ public class SoundMufflerFabric implements ClientModInitializer {
         ScreenEvents.AFTER_INIT.register((client, screen, scaledWidth, scaledHeight) ->
                 ScreenMouseEvents.afterMouseRelease(screen).register(SoundMufflerFabric::onMouseReleasePre));
         KeyBindingHelper.registerKeyBinding(soundMufflerKey);
+        Constants.isCustomSkinLoader = FabricLoader.getInstance().isModLoaded("customskinloader");
 
         //on mod keybind press
         ClientTickEvents.END_WORLD_TICK.register(level -> {