+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-Optimization changes.  
+Fixed Anchor breaking when created out of order.