+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-Optimization changes.  
+Fixed Anchors breaking when created out of order.