+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-Fixed Anchors breaking when created out of order.  
+Auto-disable Anchors when CustomSkinLoader is present, that mod breaks Read/Write their data.