+++ CHANGELOG.md
@@ -1,5 +1,4 @@
-# --ALPHA 4.01 EXPECT BUGS--
+# --ALPHA 4.02 EXPECT BUGS--
 ## Needs lots of testing! please report any bugs to the GitHub repo!
 
-- Fixed scroller not working
-- Made slight optimizations
+- Fixed anchors overriding Minecraft's sounds settings.